+++ changelog.md
@@ -218,3 +218,6 @@
 * Fixed UI Scaling issues with interaction buttons
 * Fixed issues of bounty hunters spawning within villages if your y value is below its bounds
 * Added Village Merging
+* Fixed villagers struck by lighting
+* Added electrified trait
+* Increased button widths to better support different languages